Add more features
Added Actual controller, delete post, update post, get all post comments, create comment, create follower, get by id stories, delete story, create actual, add stories in actual, get actuals by user id

+++ Anbunet.Application/Features/Posts/Delete/DeletePostCommandHandler.cs
@@ -2,6 +2,8 @@
 using Anbunet.Application.Services;
 using Anbunet.Domain.Abstractions;
 using Anbunet.Domain.Modules.Posts;
+using Microsoft.AspNetCore.Http;
+using System.Security.Claims;
 
 namespace Anbunet.Application.Features.Posts.Delete;
 
@@ -9,15 +11,21 @@ public class DeletePostCommandHandler
     (
         IPostRepository postRepository,
         IUnitOfWork unitOfWork,
-        IFileProvider fileProvider
+        IFileProvider fileProvider,
+        IHttpContextAccessor httpContextAccessor
     )
     : ICommandHandler<DeletePostCommand, Result>
 {
+    private readonly HttpContext _httpContext = httpContextAccessor.HttpContext!;
+
     public async Task<Result> Handle(DeletePostCommand request, CancellationToken cancellationToken)
     {
+        var userId = long.Parse(_httpContext.User.FindFirstValue(ClaimTypes.NameIdentifier)!);
         var post = await postRepository.GetByIdWithInclude(request.Id);
 
         if (post == null) return Result.Failure(PostErrors.PostNotFound);
+
+        if (post.UserId != userId) return Result.Failure(PostErrors.ThisIsNotPostOfThisUser);
 
         await fileProvider.Delete(post.MediaUrl);
         postRepository.Remove(post);
